Refactor tests to multi files
