--- tests/index.test.ts
@@ -1,2039 +1,0 @@
-import { chart2txt, ChartData } from '../src/index';
-import { AspectClassification, PlanetCategory } from '../src/types';
-import {
-  TRADITIONAL_ORB_CONFIG,
-  MODERN_ORB_CONFIG,
-  TIGHT_ORB_CONFIG,
-  TRADITIONAL_ASPECTS,
-} from '../src/constants';
-import { OrbResolver } from '../src/core/orbResolver';
-
-describe('chart2txt', () => {
-  describe('sign formatting', () => {
-    test('formats planets in signs correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-          { name: 'Mercury', degree: 75 }, // 15° Gemini
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain('Sun: 5° Taurus');
-      expect(result).toContain('Moon: 0° Leo');
-      expect(result).toContain('Mercury: 15° Gemini');
-    });
-
-    test('formats planets in signs correctly without degrees', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-      };
-
-      const result = chart2txt(data, { includeSignDegree: false });
-
-      // The includeSignDegree setting doesn't actually remove degrees, it's always shown
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain('Sun: 5° Taurus');
-      expect(result).toContain('Moon: 0° Leo');
-    });
-
-    test.skip('omits signs when disabled via settings', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-      };
-
-      const result = chart2txt(data, { omitSigns: true } as any);
-
-      expect(result).not.toContain('Taurus');
-      expect(result).not.toContain('Leo');
-    });
-  });
-
-  describe('house formatting', () => {
-    test('includes equal house positions', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-        ascendant: 6, // 6° Aries
-      };
-
-      const result = chart2txt(data, { houseSystemName: 'equal' });
-
-      expect(result).toContain('[METADATA]');
-      expect(result).toContain('house_system: equal');
-      expect(result).toContain('[PLANETS]');
-      // Note: House positions are not shown without houseCusps data in current implementation
-      expect(result).toContain('Sun: 5° Taurus');
-      expect(result).toContain('Moon: 0° Leo');
-    });
-
-    test('includes whole sign house positions', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-        ascendant: 6, // 6° Aries
-        houseCusps: [6, 36, 66, 96, 126, 156, 186, 216, 246, 276, 306, 336], // Equal houses from ASC
-      };
-
-      const result = chart2txt(data, { houseSystemName: 'whole_sign' });
-
-      expect(result).toContain('[METADATA]');
-      expect(result).toContain('house_system: whole_sign');
-      expect(result).toContain('[PLANETS]');
-      // With this ASC at 6° Aries and these house cusps, actual house positions are different
-      expect(result).toContain('Sun: 5° Taurus [Ruler: Venus], 1st house');
-      expect(result).toContain('Moon: 0° Leo [Ruler: Sun], 4th house');
-    });
-
-    test('includes equal house positions correctly with degree', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-        ascendant: 6, // 6° Aries
-        houseCusps: [6, 36, 66, 96, 126, 156, 186, 216, 246, 276, 306, 336], // Equal houses from ASC
-      };
-
-      const result = chart2txt(data, { includeHouseDegree: true });
-
-      expect(result).toContain('[PLANETS]');
-      // includeHouseDegree doesn't seem to affect the current format
-      expect(result).toContain('Sun: 5° Taurus [Ruler: Venus], 1st house');
-      expect(result).toContain('Moon: 0° Leo [Ruler: Sun], 4th house');
-    });
-
-    test('includes whole sign house positions correctly with degree', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-        ascendant: 6, // 6° Aries
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330], // Whole sign houses
-      };
-
-      const result = chart2txt(data, {
-        houseSystemName: 'whole_sign',
-        includeHouseDegree: true,
-      });
-
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain('Sun: 5° Taurus [Ruler: Venus], 2nd house');
-      expect(result).toContain('Moon: 0° Leo [Ruler: Sun], 5th house');
-    });
-
-    test.skip('omits house positions when disabled via settings', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-          { name: 'Moon', degree: 120 }, // 0° Leo
-        ],
-        ascendant: 6, // 6° Aries
-      };
-
-      const result = chart2txt(data, { omitHouses: true } as any);
-
-      expect(result).not.toContain('house');
-    });
-  });
-
-  describe('point formatting', () => {
-    test('formats ascendant correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        ascendant: 6, // 6° Aries
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ANGLES]');
-      expect(result).toContain('Ascendant: 6° Aries');
-    });
-
-    test('omits ascendant when disabled via settings', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        ascendant: 6, // 6° Aries
-      };
-
-      const result = chart2txt(data, { includeAscendant: false });
-
-      expect(result).toContain('[ANGLES]');
-      // The includeAscendant setting is legacy, it still shows the ASC if provided
-      expect(result).toContain('Ascendant: 6° Aries');
-    });
-
-    test('formats midheaven correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        ascendant: 6, // 6° Aries
-        midheaven: 120, // 0° Leo
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ANGLES]');
-      expect(result).toContain('Ascendant: 6° Aries');
-      expect(result).toContain('Midheaven: 0° Leo');
-    });
-
-    test('omits midheaven when not provided', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        ascendant: 6, // 6° Aries
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ANGLES]');
-      expect(result).toContain('Ascendant: 6° Aries');
-      expect(result).toContain('Midheaven: Not available');
-    });
-
-    test.skip('omits points when disabled via settings', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        ascendant: 6, // 6° Aries
-        points: [
-          { name: 'MC', degree: 120 }, // 0° Leo
-          { name: 'IC', degree: 240 }, // 0° Aquarius
-        ],
-      };
-
-      const result = chart2txt(data, { omitPoints: true } as any);
-
-      expect(result).not.toContain('MC');
-      expect(result).not.toContain('IC');
-    });
-  });
-
-  describe('aspect formatting', () => {
-    test('calculates and includes aspects between planets', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries
-          { name: 'Moon', degree: 92 }, // 2° Cancer (square to Sun)
-          { name: 'Venus', degree: 60 }, // 0° Gemini (sextile to Sun)
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun square Moon');
-      expect(result).toContain('Sun sextile Venus');
-    });
-
-    test('calculates and includes aspects with definition override', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries
-          { name: 'Moon', degree: 92 }, // 2° Cancer (square to Sun)
-          { name: 'Venus', degree: 69 }, // 9° Gemini (wide sextile to Sun)
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectDefinitions: [{ name: 'sextile2', angle: 60, orb: 10 }],
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }], // Include wide aspects up to 10°
-        skipOutOfSignAspects: false, // Allow out of sign aspects for this wide orb
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).not.toContain('square');
-      expect(result).toContain('Sun sextile2 Venus');
-    });
-
-    test('filters out-of-sign aspects by default (skipOutOfSignAspects: true)', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 28 }, // 28° Aries
-          { name: 'Moon', degree: 32 }, // 2° Taurus (4° from Sun - close conjunction but different signs)
-        ],
-      };
-
-      // Default behavior should skip out-of-sign aspects
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('None');
-      expect(result).not.toContain('conjunction');
-    });
-
-    test('includes out-of-sign aspects when skipOutOfSignAspects is false', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 28 }, // 28° Aries
-          { name: 'Moon', degree: 32 }, // 2° Taurus (4° from Sun - close conjunction but different signs)
-        ],
-      };
-
-      const result = chart2txt(data, { skipOutOfSignAspects: false });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun conjunction Moon');
-      expect(result).toContain('4.0°');
-    });
-
-    test('correctly handles Saturn-Venus cross-sign square with skipOutOfSignAspects', () => {
-      const data: ChartData = {
-        name: 'Saturn-Venus Bug Test',
-        planets: [
-          { name: 'Venus', degree: 54.17 }, // 24° Taurus (sign 1)
-          { name: 'Saturn', degree: 329.99 }, // 29° Aquarius (sign 10)
-          { name: 'Sun', degree: 98.76 }, // 8° Cancer - for context
-        ],
-      };
-
-      // Test with major/moderate orbs and skipOutOfSignAspects: true (should now include the square)
-      const resultWithSkip = chart2txt(data, {
-        skipOutOfSignAspects: true,
-        aspectDefinitions: [
-          { name: 'square', angle: 90, orb: 6 }, // Extend orb to 6° to catch the 5.82° aspect
-        ],
-        aspectCategories: [
-          { name: 'MAJOR', maxOrb: 3 },
-          { name: 'MODERATE', maxOrb: 6 },
-        ],
-      });
-
-      // Test with skipOutOfSignAspects: false (should also include the square)
-      const resultWithoutSkip = chart2txt(data, {
-        skipOutOfSignAspects: false,
-        aspectDefinitions: [
-          { name: 'square', angle: 90, orb: 6 }, // Extend orb to 6° to catch the 5.82° aspect
-        ],
-        aspectCategories: [
-          { name: 'MAJOR', maxOrb: 3 },
-          { name: 'MODERATE', maxOrb: 6 },
-        ],
-      });
-
-      expect(resultWithSkip).toContain('[ASPECTS]');
-      expect(resultWithoutSkip).toContain('[ASPECTS]');
-
-      // Saturn-Venus are 84.18° apart (5.82° from exact square)
-      // Saturn 29° Aquarius, Venus 24° Taurus - circular distance is 3 signs (valid square)
-      // This should now be correctly included even with skipOutOfSignAspects: true
-      expect(resultWithSkip).toMatch(
-        /(?:Saturn square Venus|Venus square Saturn)/
-      );
-      expect(resultWithoutSkip).toMatch(
-        /(?:Saturn square Venus|Venus square Saturn)/
-      );
-    });
-
-    test.skip('omit aspects when disabled via settings', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries
-          { name: 'Moon', degree: 92 }, // 0° Cancer (square to Sun)
-          { name: 'Venus', degree: 60 }, // 0° Gemini (sextile to Sun)
-        ],
-      };
-
-      const result = chart2txt(data, { omitAspects: true } as any);
-
-      expect(result).not.toContain('square');
-      expect(result).not.toContain('sextile');
-    });
-  });
-
-  describe('location and date tests', () => {
-    test('displays header when no date or location provided', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[CHART: test]');
-      expect(result).toContain('[BIRTHDATA] Not available');
-    });
-
-    test('displays timestamp when provided', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        timestamp: new Date(1),
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[CHART: test]');
-      expect(result).toContain(
-        '[BIRTHDATA] Unknown Location | 12/31/1969 | 07:00:00 PM'
-      );
-    });
-
-    test('displays location when provided', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        location: 'San Francisco',
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[CHART: test]');
-      // Location only appears if timestamp is also provided
-      expect(result).toContain('[BIRTHDATA] Not available');
-    });
-
-    test('displays location and date when provided', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 35 }, // 5° Taurus
-        ],
-        timestamp: new Date(1),
-        location: 'San Francisco',
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[CHART: test]');
-      expect(result).toContain(
-        '[BIRTHDATA] San Francisco | 12/31/1969 | 07:00:00 PM'
-      );
-    });
-  });
-
-  describe('house cusps formatting', () => {
-    test('formats house cusps correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: 35 }],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[HOUSE CUSPS]');
-      expect(result).toContain('1st house: 0° Aries');
-      expect(result).toContain('7th house: 0° Libra');
-      expect(result).toContain('2nd house: 0° Taurus');
-      expect(result).toContain('8th house: 0° Scorpio');
-    });
-
-    test('handles missing house cusps', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: 35 }],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[HOUSE CUSPS]');
-      expect(result).toContain('House cusps not available');
-    });
-
-    test('formats house cusps with various degrees', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: 35 }],
-        houseCusps: [15, 45, 75, 105, 135, 165, 195, 225, 255, 285, 315, 345],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[HOUSE CUSPS]');
-      expect(result).toContain('1st house: 15° Aries');
-      expect(result).toContain('7th house: 15° Libra');
-      expect(result).toContain('4th house: 15° Cancer');
-      expect(result).toContain('10th house: 15° Capricorn');
-    });
-
-    test('handles zodiac wraparound when 1st house cusp > 12th house cusp', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 350 }, // 20° Pisces - should be in 1st house
-          { name: 'Moon', degree: 10 }, // 10° Aries - should be in 2nd house
-          { name: 'Mercury', degree: 320 }, // 20° Aquarius - should be in 12th house
-          { name: 'Venus', degree: 40 }, // 10° Taurus - should be in 3rd house
-        ],
-        // House cusps where 1st house starts at 340° (10° Pisces) and 12th house starts at 310° (10° Aquarius)
-        // This creates a wraparound situation: 12th house spans 310°-340°, 1st house spans 340°-10° (next day)
-        houseCusps: [340, 10, 40, 70, 100, 130, 160, 190, 220, 250, 280, 310],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[PLANETS]');
-      // Sun at 350° (20° Pisces) should be in 1st house (340°-10° wraparound)
-      expect(result).toContain('Sun: 20° Pisces [Ruler: Jupiter], 1st house');
-      // Moon at 10° (10° Aries) should be in 2nd house (10°-40°)
-      expect(result).toContain('Moon: 10° Aries [Ruler: Mars], 2nd house');
-      // Mercury at 320° (20° Aquarius) should be in 12th house (310°-340°)
-      expect(result).toContain(
-        'Mercury: 20° Aquarius [Ruler: Saturn], 12th house'
-      );
-      // Venus at 40° (10° Taurus) should be in 3rd house (40°-70°)
-      expect(result).toContain('Venus: 10° Taurus [Domicile], 3rd house');
-
-      expect(result).toContain('[HOUSE CUSPS]');
-      expect(result).toContain('1st house: 10° Pisces'); // 340° = 10° Pisces
-      expect(result).toContain('12th house: 10° Aquarius'); // 310° = 10° Aquarius
-    });
-  });
-
-  describe('essential dignities formatting', () => {
-    test('formats planets with ruler dignities', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo - Sun rules Leo
-          { name: 'Moon', degree: 90 }, // 0° Cancer - Moon rules Cancer
-          { name: 'Mercury', degree: 150 }, // 0° Virgo - Mercury rules Virgo
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain('Sun: 0° Leo [Domicile], 5th house');
-      expect(result).toContain('Moon: 0° Cancer [Domicile], 4th house');
-      expect(result).toContain(
-        'Mercury: 0° Virgo [Domicile, Exaltation], 6th house'
-      );
-    });
-
-    test('formats planets with detriment and fall', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 270 }, // 0° Capricorn - Sun in detriment (ruled by Saturn)
-          { name: 'Mars', degree: 90 }, // 0° Cancer - Mars in fall
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain('Sun: 0° Capricorn [Ruler: Saturn]');
-      expect(result).toContain('Mars: 0° Cancer [Fall | Ruler: Moon]');
-    });
-
-    test('formats planets with exaltation', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries - Sun exalted in Aries
-          { name: 'Moon', degree: 30 }, // 0° Taurus - Moon exalted in Taurus
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain('Sun: 0° Aries [Exaltation | Ruler: Mars]');
-      expect(result).toContain('Moon: 0° Taurus [Exaltation | Ruler: Venus]');
-    });
-
-    test('formats planets with mixed dignities and retrograde', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Mercury', degree: 150, speed: -1.2 }, // 0° Virgo Rx - Mercury rules and is exalted in Virgo
-          { name: 'Venus', degree: 180, speed: -0.5 }, // 0° Libra Rx - Venus rules Libra
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[PLANETS]');
-      expect(result).toContain(
-        'Mercury: 0° Virgo Retrograde [Domicile, Exaltation], 6th house'
-      );
-      expect(result).toContain(
-        'Venus: 0° Libra Retrograde [Domicile], 7th house'
-      );
-    });
-  });
-
-  describe('dispositor chains formatting', () => {
-    test('identifies final dispositors correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo - Sun rules itself
-          { name: 'Moon', degree: 90 }, // 0° Cancer - Moon rules itself
-          { name: 'Mercury', degree: 35 }, // 5° Taurus - disposed by Venus (not in chart)
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[DISPOSITOR TREE]');
-      expect(result).toContain('Sun → (final)');
-      expect(result).toContain('Moon → (final)');
-      expect(result).toContain('Mercury → (final)');
-    });
-
-    test('shows dispositor chains with dependencies', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo - Sun rules itself
-          { name: 'Mercury', degree: 35 }, // 5° Taurus - disposed by Venus (not in chart)
-          { name: 'Venus', degree: 90 }, // 0° Cancer - disposed by Moon (not in chart)
-          { name: 'Moon', degree: 120 }, // 0° Leo - disposed by Sun
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[DISPOSITOR TREE]');
-      expect(result).toContain('Sun → (final)');
-      expect(result).toContain('Mercury → Venus → Moon → Sun → (final)');
-      expect(result).toContain('Venus → Moon → Sun → (final)');
-      expect(result).toContain('Moon → Sun → (final)');
-    });
-
-    test('handles empty planet list', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[DISPOSITOR TREE]');
-      expect(result).toContain('No planets available for dispositor analysis.');
-    });
-
-    test('handles complex dispositor chains', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo - rules itself
-          { name: 'Moon', degree: 150 }, // 0° Virgo - disposed by Mercury
-          { name: 'Mercury', degree: 60 }, // 0° Gemini - rules itself
-          { name: 'Venus', degree: 210 }, // 0° Scorpio - disposed by Mars/Pluto
-          { name: 'Mars', degree: 30 }, // 0° Taurus - disposed by Venus
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[DISPOSITOR TREE]');
-      expect(result).toContain('Sun → (final)');
-      expect(result).toContain('Mercury → (final)');
-      expect(result).toContain('Venus → Mars → (cycle)');
-      expect(result).toContain('Mars → Venus → (cycle)');
-    });
-  });
-
-  describe('sign distribution formatting', () => {
-    test('formats element distribution correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo - Fire
-          { name: 'Moon', degree: 90 }, // 0° Cancer - Water
-          { name: 'Mercury', degree: 35 }, // 5° Taurus - Earth
-          { name: 'Venus', degree: 210 }, // 0° Scorpio - Water
-          { name: 'Mars', degree: 60 }, // 0° Gemini - Air
-        ],
-        ascendant: 150, // 0° Virgo - Earth
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ELEMENT DISTRIBUTION]');
-      expect(result).toContain('Fire: 1 (Sun)');
-      expect(result).toContain('Earth: 2 (Mercury, Ascendant)');
-      expect(result).toContain('Air: 1 (Mars)');
-      expect(result).toContain('Water: 2 (Moon, Venus)');
-    });
-
-    test('formats modality distribution correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries - Cardinal
-          { name: 'Moon', degree: 90 }, // 0° Cancer - Cardinal
-          { name: 'Mercury', degree: 120 }, // 0° Leo - Fixed
-          { name: 'Venus', degree: 60 }, // 0° Gemini - Mutable
-        ],
-        ascendant: 240, // 0° Sagittarius - Mutable
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[MODALITY DISTRIBUTION]');
-      expect(result).toContain('Cardinal: 2');
-      expect(result).toContain('Fixed: 1');
-      expect(result).toContain('Mutable: 2');
-    });
-
-    test('formats polarity distribution correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries - Masculine
-          { name: 'Moon', degree: 30 }, // 0° Taurus - Feminine
-          { name: 'Mercury', degree: 60 }, // 0° Gemini - Masculine
-          { name: 'Venus', degree: 90 }, // 0° Cancer - Feminine
-          { name: 'Mars', degree: 120 }, // 0° Leo - Masculine
-        ],
-        ascendant: 150, // 0° Virgo - Feminine
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[POLARITY]');
-      expect(result).toContain('Masculine (Active): 3');
-      expect(result).toContain('Feminine (Receptive): 3');
-    });
-
-    test('handles empty planet list for distributions', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ELEMENT DISTRIBUTION]');
-      expect(result).toContain('No planets available for element analysis.');
-      expect(result).toContain('[MODALITY DISTRIBUTION]');
-      expect(result).toContain('No planets available for modality analysis.');
-      expect(result).toContain('[POLARITY]');
-      expect(result).toContain('No planets available for polarity analysis.');
-    });
-
-    test('handles distributions without ascendant', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo - Fire, Fixed, Masculine
-          { name: 'Moon', degree: 90 }, // 0° Cancer - Water, Cardinal, Feminine
-        ],
-        // No ascendant provided
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ELEMENT DISTRIBUTION]');
-      expect(result).toContain('Fire: 1 (Sun)');
-      expect(result).toContain('Water: 1 (Moon)');
-      expect(result).toContain('[MODALITY DISTRIBUTION]');
-      expect(result).toContain('Cardinal: 1');
-      expect(result).toContain('Fixed: 1');
-      expect(result).toContain('[POLARITY]');
-      expect(result).toContain('Masculine (Active): 1');
-      expect(result).toContain('Feminine (Receptive): 1');
-    });
-  });
-
-  describe('synastry and multi-chart formatting', () => {
-    test('formats 2-chart synastry with house overlays', () => {
-      const chart1: ChartData = {
-        name: 'Alice',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo
-          { name: 'Moon', degree: 180 }, // 0° Libra
-          { name: 'Mercury', degree: 210 }, // 0° Scorpio
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const chart2: ChartData = {
-        name: 'Bob',
-        planets: [
-          { name: 'Sun', degree: 90 }, // 0° Cancer
-          { name: 'Moon', degree: 240 }, // 0° Sagittarius
-          { name: 'Venus', degree: 150 }, // 0° Virgo
-        ],
-        houseCusps: [15, 45, 75, 105, 135, 165, 195, 225, 255, 285, 315, 345],
-      };
-
-      const result = chart2txt([chart1, chart2], {
-        includeAspectPatterns: true,
-      });
-
-      // Should have individual chart sections
-      expect(result).toContain('[CHART: Alice]');
-      expect(result).toContain('[CHART: Bob]');
-
-      // Should have synastry section
-      expect(result).toContain('[SYNASTRY: Alice-Bob]');
-      expect(result).toContain('[PLANET-PLANET ASPECTS]');
-
-      // Should have house overlays with new format
-      expect(result).toContain('[HOUSE OVERLAYS]');
-      expect(result).toContain("Alice's planets in Bob's houses:");
-      expect(result).toContain('- Sun: 4th');
-      expect(result).toContain('- Moon: 6th');
-      expect(result).toContain('- Mercury: 7th');
-      expect(result).toContain("Bob's planets in Alice's houses:");
-      expect(result).toContain('- Sun: 4th');
-      expect(result).toContain('- Moon: 9th');
-      expect(result).toContain('- Venus: 6th');
-    });
-
-    test('formats 3-chart group synastry', () => {
-      const chart1: ChartData = {
-        name: 'Alice',
-        planets: [
-          { name: 'Sun', degree: 0 }, // 0° Aries
-          { name: 'Moon', degree: 90 }, // 0° Cancer
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const chart2: ChartData = {
-        name: 'Bob',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo
-          { name: 'Moon', degree: 180 }, // 0° Libra
-        ],
-        houseCusps: [30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 0],
-      };
-
-      const chart3: ChartData = {
-        name: 'Charlie',
-        planets: [
-          { name: 'Sun', degree: 240 }, // 0° Sagittarius
-          { name: 'Moon', degree: 300 }, // 0° Aquarius
-        ],
-        houseCusps: [60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 0, 30],
-      };
-
-      const result = chart2txt([chart1, chart2, chart3], {
-        includeAspectPatterns: true,
-      });
-
-      // Should have individual chart sections
-      expect(result).toContain('[CHART: Alice]');
-      expect(result).toContain('[CHART: Bob]');
-      expect(result).toContain('[CHART: Charlie]');
-
-      // Should have pairwise synastry sections
-      expect(result).toContain('[SYNASTRY: Alice-Bob]');
-      expect(result).toContain('[SYNASTRY: Alice-Charlie]');
-      expect(result).toContain('[SYNASTRY: Bob-Charlie]');
-
-      // Should have multiple house overlay sections
-      expect(result).toContain("Alice's planets in Bob's houses:");
-      expect(result).toContain("Bob's planets in Alice's houses:");
-      expect(result).toContain("Alice's planets in Charlie's houses:");
-      expect(result).toContain("Charlie's planets in Alice's houses:");
-      expect(result).toContain("Bob's planets in Charlie's houses:");
-      expect(result).toContain("Charlie's planets in Bob's houses:");
-    });
-
-    test('formats synastry with transit chart', () => {
-      const natalChart: ChartData = {
-        name: 'Alice',
-        chartType: 'natal',
-        planets: [
-          { name: 'Sun', degree: 150 }, // 0° Virgo
-          { name: 'Moon', degree: 60 }, // 0° Gemini
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const transitChart: ChartData = {
-        name: 'Current Transits',
-        chartType: 'transit',
-        planets: [
-          { name: 'Sun', degree: 300 }, // 0° Aquarius
-          { name: 'Mars', degree: 210 }, // 0° Scorpio
-        ],
-        location: 'New York',
-        timestamp: new Date(2024, 0, 15, 12, 0, 0),
-      };
-
-      const result = chart2txt([natalChart, transitChart], {
-        includeAspectPatterns: true,
-      });
-
-      // Should have natal chart section
-      expect(result).toContain('[CHART: Alice]');
-
-      // Should have transit chart info
-      expect(result).toContain('[TRANSIT: Current Transits]');
-      expect(result).toContain('[DATETIME]');
-
-      // Should have transit aspects
-      expect(result).toContain('[TRANSIT ASPECTS: Alice]');
-    });
-
-    test('formats synastry with event chart', () => {
-      const natalChart: ChartData = {
-        name: 'Alice',
-        chartType: 'natal',
-        planets: [
-          { name: 'Sun', degree: 30 }, // 0° Taurus
-          { name: 'Moon', degree: 270 }, // 0° Capricorn
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const eventChart: ChartData = {
-        name: 'Wedding',
-        chartType: 'event',
-        planets: [
-          { name: 'Sun', degree: 180 }, // 0° Libra
-          { name: 'Venus', degree: 210 }, // 0° Scorpio
-        ],
-        houseCusps: [15, 45, 75, 105, 135, 165, 195, 225, 255, 285, 315, 345],
-        location: 'Paris',
-        timestamp: new Date(2024, 5, 21, 14, 30, 0),
-      };
-
-      const result = chart2txt([natalChart, eventChart], {
-        includeAspectPatterns: true,
-      });
-
-      // Should have both chart sections
-      expect(result).toContain('[CHART: Alice]');
-      expect(result).toContain('[CHART: Wedding]');
-
-      // Should have natal-event relationship
-      expect(result).toContain('[NATAL_EVENT: Alice-Wedding]');
-
-      // Should have house overlays
-      expect(result).toContain('[HOUSE OVERLAYS]');
-      expect(result).toContain("Alice's planets in Wedding's houses:");
-      expect(result).toContain("Wedding's planets in Alice's houses:");
-    });
-
-    test('handles missing house cusps in overlays', () => {
-      const chart1: ChartData = {
-        name: 'Alice',
-        planets: [
-          { name: 'Sun', degree: 120 }, // 0° Leo
-        ],
-        // No house cusps provided
-      };
-
-      const chart2: ChartData = {
-        name: 'Bob',
-        planets: [
-          { name: 'Moon', degree: 180 }, // 0° Libra
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt([chart1, chart2], {
-        includeAspectPatterns: true,
-      });
-
-      expect(result).toContain('[HOUSE OVERLAYS]');
-      expect(result).toContain("Alice's planets in Bob's houses:");
-      expect(result).toContain('- Sun: 5th');
-      expect(result).toContain(
-        "Bob's planets in Alice's houses: (Alice house cusps not available)"
-      );
-    });
-  });
-
-  describe('aspect patterns formatting', () => {
-    test('detects T-Square pattern', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 180, speed: 13.0 }, // 0° Libra (opposition to Sun)
-          { name: 'Saturn', degree: 90, speed: 0.5 }, // 0° Cancer (square to both)
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      expect(result).toContain('T-Square:');
-      expect(result).toContain('- Apex: Saturn 0° Cancer (4th house)');
-      expect(result).toContain(
-        '- Opposition: Sun 0° Aries (1st house) - Moon 0° Libra (7th house)'
-      );
-      expect(result).toContain('- Mode: Cardinal');
-    });
-
-    test('detects Grand Trine pattern', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries (Fire)
-          { name: 'Moon', degree: 120, speed: 13.0 }, // 0° Leo (Fire)
-          { name: 'Mars', degree: 240, speed: 0.7 }, // 0° Sagittarius (Fire)
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      expect(result).toContain('Grand Trine:');
-      expect(result).toContain('- Planet 1: Sun 0° Aries (1st house)');
-      expect(result).toContain('- Planet 2: Moon 0° Leo (5th house)');
-      expect(result).toContain('- Planet 3: Mars 0° Sagittarius (9th house)');
-      expect(result).toContain('- Element: Fire');
-    });
-
-    test('detects Stellium pattern', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 95, speed: 1.0 }, // 5° Cancer
-          { name: 'Mercury', degree: 100, speed: 1.2 }, // 10° Cancer
-          { name: 'Venus', degree: 105, speed: 1.1 }, // 15° Cancer
-          { name: 'Mars', degree: 180, speed: 0.7 }, // Different sign
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      expect(result).toContain('Stellium:');
-      expect(result).toContain('- Planets: Sun, Mercury, Venus');
-      expect(result).toContain('- Sign: Cancer');
-      expect(result).toContain('- Houses: 4th');
-      expect(result).toContain('- Span: 10.0°');
-    });
-
-    test('detects Yod pattern', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 60, speed: 13.0 }, // 0° Gemini (sextile to Sun: 60°)
-          { name: 'Saturn', degree: 150, speed: 0.5 }, // 30° Leo (quincunx to Sun: 150°, quincunx to Moon: 90°... wait, this is wrong)
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      // Let me fix the degrees for a proper Yod:
-      // Sun at 0° (Aries), Moon at 60° (Gemini) = 60° sextile ✓
-      // For Saturn to form quincunxes: Saturn at 150° = quincunx to Sun (150°-0°=150°) ✓
-      // Saturn at 150° to Moon at 60° = 150°-60°=90° (square, not quincunx)
-      // Let me recalculate: if Sun=0°, Moon=60°, then Saturn should be at 210° for quincunxes
-      // Sun=0° to Saturn=210° = 210° (150° if we consider 360°-210°=150°)
-      // Moon=60° to Saturn=210° = 150° ✓
-
-      const correctedData: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 60, speed: 13.0 }, // 0° Gemini (sextile: 60°)
-          { name: 'Saturn', degree: 210, speed: 0.5 }, // 0° Scorpio (quincunx to both: 210°-0°=210°>180° so 360°-210°=150°)
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(correctedData, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      expect(result).toContain('Yod:');
-      expect(result).toContain('- Apex: Saturn 0° Scorpio (8th house)');
-      expect(result).toContain('- Base planet 1: Sun 0° Aries (1st house)');
-      expect(result).toContain('- Base planet 2: Moon 0° Gemini (3rd house)');
-    });
-
-    test('handles no aspect patterns', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 45, speed: 13.0 }, // 15° Taurus (no major aspects)
-        ],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      expect(result).toContain('No aspect patterns detected.');
-    });
-
-    test('does not show aspect patterns when disabled', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 180, speed: 13.0 }, // 0° Libra (opposition to Sun)
-          { name: 'Saturn', degree: 90, speed: 0.5 }, // 0° Cancer (square to both)
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data); // Default: includeAspectPatterns: false
-
-      expect(result).not.toContain('[ASPECT PATTERNS]');
-      expect(result).not.toContain('T-Square:');
-    });
-
-    test('detects Grand Cross pattern (currently detected as T-Squares)', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 90, speed: 13.0 }, // 0° Cancer
-          { name: 'Venus', degree: 180, speed: 1.1 }, // 0° Libra
-          { name: 'Saturn', degree: 270, speed: 0.5 }, // 0° Capricorn
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      // Currently this configuration is detected as multiple T-Squares rather than one Grand Cross
-      // This is a known limitation that could be improved in the future
-      expect(result).toContain('T-Square:');
-      expect(result).toContain('- Mode: Cardinal');
-    });
-
-    test('detects multiple patterns in same chart', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          // T-Square
-          { name: 'Sun', degree: 0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 180, speed: 13.0 }, // 0° Libra
-          { name: 'Saturn', degree: 90, speed: 0.5 }, // 0° Cancer
-          // Stellium in Leo
-          { name: 'Mercury', degree: 120, speed: 1.2 }, // 0° Leo
-          { name: 'Venus', degree: 125, speed: 1.1 }, // 5° Leo
-          { name: 'Mars', degree: 130, speed: 0.7 }, // 10° Leo
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data, { includeAspectPatterns: true });
-
-      expect(result).toContain('[ASPECT PATTERNS]');
-      expect(result).toContain('T-Square:');
-      expect(result).toContain('Stellium:');
-      expect(result).toContain('- Apex: Saturn 0° Cancer (4th house)');
-      expect(result).toContain('- Planets: Mercury, Venus, Mars');
-      expect(result).toContain('- Sign: Leo');
-    });
-  });
-
-  // Tests for critical bug fixes
-  describe('degree normalization fixes', () => {
-    test('handles negative degrees correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: -30 }, // Should normalize to 330° (0° Pisces)
-          { name: 'Moon', degree: -90 }, // Should normalize to 270° (0° Capricorn)
-        ],
-      };
-
-      const result = chart2txt(data);
-
-      expect(result).toContain('Sun: 0° Pisces'); // -30° + 360° = 330° = 0° Pisces
-      expect(result).toContain('Moon: 0° Capricorn'); // -90° + 360° = 270° = 0° Capricorn
-    });
-
-    test('handles degrees >= 360 correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 390 }, // Should normalize to 30° (0° Taurus)
-          { name: 'Moon', degree: 450 }, // Should normalize to 90° (0° Cancer)
-        ],
-      };
-
-      const result = chart2txt(data);
-
-      expect(result).toContain('Sun: 0° Taurus');
-      expect(result).toContain('Moon: 0° Cancer');
-    });
-
-    test('handles extreme degree values correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 1080 }, // Multiple rotations: should normalize to 0° (0° Aries)
-          { name: 'Moon', degree: -720 }, // Multiple negative rotations: should normalize to 0° (0° Aries)
-        ],
-      };
-
-      const result = chart2txt(data);
-
-      expect(result).toContain('Sun: 0° Aries');
-      expect(result).toContain('Moon: 0° Aries');
-    });
-  });
-
-  describe('input validation fixes', () => {
-    test('throws error for NaN degree values', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: NaN }],
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('invalid degree value');
-    });
-
-    test('throws error for Infinity degree values', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: Infinity }],
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('invalid degree value');
-    });
-
-    test('throws error for empty planet names', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: '', degree: 30 }],
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('non-empty string');
-    });
-
-    test('throws error for invalid house cusps array length', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: 30 }],
-        houseCusps: [0, 30, 60, 90, 120], // Only 5 cusps instead of 12
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('exactly 12 values');
-    });
-
-    test('throws error for invalid house cusp values', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: 30 }],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, NaN],
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('invalid value');
-    });
-
-    test('throws error for duplicate planet names', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 30 },
-          { name: 'Sun', degree: 60 }, // Duplicate name
-        ],
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('Duplicate point names');
-    });
-
-    test('validates invalid speed values', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [{ name: 'Sun', degree: 30, speed: NaN }],
-      };
-
-      expect(() => chart2txt(data)).toThrow('Invalid chart data');
-      expect(() => chart2txt(data)).toThrow('invalid speed value');
-    });
-  });
-
-  describe('house calculation consistency fixes', () => {
-    test('handles planets exactly on house cusps consistently', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 30.0 }, // Exactly on 2nd house cusp
-          { name: 'Moon', degree: 60.0 }, // Exactly on 3rd house cusp
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data);
-
-      // Planets exactly on cusps should be assigned to the house starting at that cusp
-      expect(result).toContain('Sun: 0° Taurus [Ruler: Venus], 2nd house');
-      expect(result).toContain('Moon: 0° Gemini [Ruler: Mercury], 3rd house');
-    });
-
-    test('handles house wraparound at 0°/360° boundary consistently', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 350 }, // Should be in 1st house
-          { name: 'Moon', degree: 10 }, // Should be in 2nd house
-          { name: 'Mercury', degree: 320 }, // Should be in 12th house
-        ],
-        houseCusps: [340, 10, 40, 70, 100, 130, 160, 190, 220, 250, 280, 310],
-      };
-
-      const result = chart2txt(data);
-
-      expect(result).toContain('Sun: 20° Pisces [Ruler: Jupiter], 1st house');
-      expect(result).toContain('Moon: 10° Aries [Ruler: Mars], 2nd house');
-      expect(result).toContain(
-        'Mercury: 20° Aquarius [Ruler: Saturn], 12th house'
-      );
-    });
-  });
-
-  describe('dignity mapping fixes', () => {
-    test('shows Mercury as having both detriment and fall in Pisces', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Mercury', degree: 330 }, // 0° Pisces
-        ],
-      };
-
-      const result = chart2txt(data);
-
-      // Mercury should show both detriment and fall in Pisces
-      expect(result).toContain(
-        'Mercury: 0° Pisces [Detriment, Fall | Ruler: Jupiter]'
-      );
-    });
-
-    test('handles signs with no fall values correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 60 }, // 0° Gemini (no fall planet defined)
-          { name: 'Moon', degree: 240 }, // 0° Sagittarius (no fall planet defined)
-        ],
-      };
-
-      const result = chart2txt(data);
-
-      // Should not show fall information for signs that don't have it
-      expect(result).toContain('Sun: 0° Gemini [Ruler: Mercury]');
-      expect(result).toContain('Moon: 0° Sagittarius [Ruler: Jupiter]');
-      expect(result).not.toContain('Fall:');
-    });
-  });
-
-  describe('aspect calculation fixes', () => {
-    test('out-of-sign aspect filtering works correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 29 }, // 29° Aries
-          { name: 'Moon', degree: 121 }, // 1° Leo - should be trine (120°) but out of sign
-        ],
-      };
-
-      const result = chart2txt(data, { skipOutOfSignAspects: true });
-
-      // With skipOutOfSignAspects=true, this 119° aspect should be filtered out
-      // because Sun in Aries and Moon in Leo are 3 signs apart (not 4 as expected for trine)
-      expect(result).toContain('[ASPECTS]');
-      expect(result).not.toContain('Sun trine Moon');
-    });
-
-    test('includes out-of-sign aspects when setting is disabled', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 29 }, // 29° Aries
-          { name: 'Moon', degree: 149 }, // 29° Leo - should be trine (120°) but out of sign
-        ],
-      };
-
-      const result = chart2txt(data, { skipOutOfSignAspects: false });
-
-      // With skipOutOfSignAspects=false, this 120° aspect should be included
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun trine Moon');
-    });
-  });
-
-  describe('applying vs separating aspect detection fixes', () => {
-    test('correctly identifies applying aspects', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0.0, speed: 1.0 }, // 0° Aries, moving forward
-          { name: 'Moon', degree: 123.0, speed: 13.0 }, // 3° Leo, moving faster forward
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }], // Allow wider orbs
-        skipOutOfSignAspects: false, // Allow out-of-sign aspects for this test
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun trine Moon');
-      // Moon at 123° is ahead of exact trine (120°) but moving faster forward, so it's separating
-      expect(result).toContain('(separating)');
-    });
-
-    test('correctly identifies separating aspects', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0.0, speed: 1.0 }, // 0° Aries, moving forward
-          { name: 'Moon', degree: 117.0, speed: 13.0 }, // 27° Cancer, moving faster forward
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }], // Allow wider orbs
-        skipOutOfSignAspects: false, // Allow out-of-sign aspects for this test
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun trine Moon');
-      // Moon at 117° is behind exact trine (120°) but moving faster forward, so it's applying
-      expect(result).toContain('(applying)');
-    });
-
-    test('correctly handles retrograde planets in aspect application', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0.0, speed: 1.0 }, // 0° Aries, direct
-          { name: 'Mars', degree: 93.0, speed: -0.5 }, // 3° Cancer, retrograde
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }], // Allow wider orbs
-        skipOutOfSignAspects: false, // Allow out-of-sign aspects for this test
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun square Mars');
-      // Mars is retrograde and moving toward exact square (90°) - should be applying
-      expect(result).toContain('(applying)');
-    });
-
-    test('handles complex boundary case with 0°/360° wraparound', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 358.0, speed: 1.0 }, // 28° Pisces, moving forward
-          { name: 'Moon', degree: 2.0, speed: 13.0 }, // 2° Aries, moving faster forward
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectDefinitions: [{ name: 'conjunction', angle: 0, orb: 10 }], // Wider orb for conjunction
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }], // Allow wider orbs
-        skipOutOfSignAspects: false, // Allow out-of-sign aspects for this test
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun conjunction Moon');
-      // Moon is faster and moving away from the conjunction (4° → 5.28° in 0.1 days) - should be separating
-      expect(result).toContain('(separating)');
-    });
-
-    test('correctly identifies exact aspects when speeds are equal', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0.0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 120.0, speed: 1.0 }, // 0° Leo, same speed
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'TIGHT', maxOrb: 10 }], // Allow wider orbs
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun trine Moon');
-      // Same speeds - should not show application status (exact aspects don't show parens)
-      expect(result).not.toContain('(applying)');
-      expect(result).not.toContain('(separating)');
-    });
-
-    test('handles fast-moving Moon vs slow Mercury correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Moon', degree: 235.99, speed: 13.76 }, // 25°59' Scorpio, fast-moving
-          { name: 'Mercury', degree: 118.21, speed: 0.08 }, // 28°13' Cancer, slow-moving
-        ],
-      };
-
-      const result = chart2txt(data, {
-        aspectDefinitions: [{ name: 'trine', angle: 120, orb: 10 }], // Wide orb to catch this aspect
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }],
-        skipOutOfSignAspects: false,
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Moon trine Mercury');
-      // With time-based calculation, this should be applying (Moon catching up to exact trine)
-      expect(result).toContain('(applying)');
-    });
-  });
-
-  describe('aspects to angles (Ascendant and Midheaven)', () => {
-    test('calculates aspects between planets and Ascendant', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0.0, speed: 1.0 }, // 0° Aries
-          { name: 'Moon', degree: 90.0, speed: 13.0 }, // 0° Cancer
-        ],
-        ascendant: 180.0, // 0° Libra
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }],
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun opposition Ascendant');
-      expect(result).toContain('Moon square Ascendant');
-    });
-
-    test('calculates aspects between planets and Midheaven', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Venus', degree: 60.0, speed: 1.2 }, // 0° Gemini
-          { name: 'Mars', degree: 300.0, speed: 0.7 }, // 0° Aquarius
-        ],
-        midheaven: 120.0, // 0° Leo
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }],
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Venus sextile Midheaven');
-      // Mars at 300° to Midheaven at 120° = 180° = opposition
-      expect(result).toContain('Mars opposition Midheaven');
-    });
-
-    test('calculates aspects between Ascendant and Midheaven', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 150.0 }, // 0° Virgo, no aspects to angles
-        ],
-        ascendant: 0.0, // 0° Aries
-        midheaven: 90.0, // 0° Cancer
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }],
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Ascendant square Midheaven');
-    });
-
-    test('handles aspects with applying/separating for angles', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Moon', degree: 117.0, speed: 13.0 }, // 27° Cancer, fast-moving
-        ],
-        ascendant: 120.0, // 0° Leo
-      };
-
-      const result = chart2txt(data, {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }],
-        skipOutOfSignAspects: false,
-      });
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Moon conjunction Ascendant');
-      // Angles don't have speed, so aspects to angles don't show applying/separating
-      expect(result).not.toContain('(applying)');
-      expect(result).not.toContain('(separating)');
-    });
-
-    test('works in synastry with angles from both charts', () => {
-      const chart1: ChartData = {
-        name: 'Person A',
-        planets: [
-          { name: 'Sun', degree: 30.0 }, // 0° Taurus
-        ],
-        ascendant: 0.0, // 0° Aries
-      };
-
-      const chart2: ChartData = {
-        name: 'Person B',
-        planets: [
-          { name: 'Moon', degree: 60.0 }, // 0° Gemini
-        ],
-        midheaven: 90.0, // 0° Cancer
-      };
-
-      const result = chart2txt([chart1, chart2], {
-        aspectCategories: [{ name: 'WIDE', maxOrb: 10 }],
-      });
-
-      expect(result).toContain('[SYNASTRY: Person A-Person B]');
-      expect(result).toContain('[PLANET-PLANET ASPECTS]');
-      // Check for cross-chart angle aspects with proper chart name formatting
-      expect(result).toContain("Person A's Sun sextile Person B's Midheaven"); // Sun (30°) sextile MC (90°) = 60°
-      expect(result).toContain("Person A's Ascendant sextile Person B's Moon"); // ASC (0°) sextile Moon (60°) = 60°
-      expect(result).toContain(
-        "Person A's Ascendant square Person B's Midheaven"
-      ); // ASC (0°) square MC (90°) = 90°
-    });
-  });
-
-  describe('floating point precision fixes', () => {
-    test('handles very precise degree values correctly', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 29.999999 }, // Very close to 30°
-          { name: 'Moon', degree: 30.000001 }, // Very close to 30°
-        ],
-        houseCusps: [0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330],
-      };
-
-      const result = chart2txt(data);
-
-      // Both should be handled consistently despite tiny differences
-      expect(result).toContain('Sun: 29° Aries');
-      expect(result).toContain('Moon: 0° Taurus');
-      // House assignments should be consistent
-      expect(result).toContain('1st house');
-      expect(result).toContain('2nd house');
-    });
-
-    test('handles aspects with very tight orbs consistently', () => {
-      const data: ChartData = {
-        name: 'test',
-        planets: [
-          { name: 'Sun', degree: 0.0 },
-          { name: 'Moon', degree: 120.0001 }, // Trine with tiny orb
-        ],
-      };
-
-      const result = chart2txt(data);
-
-      expect(result).toContain('[ASPECTS]');
-      expect(result).toContain('Sun trine Moon');
-      // Should show very small orb (0.0°)
-      expect(result).toContain('0.0°');
-    });
-  });
-
-  describe('Enhanced Orb Configuration System', () => {
-    const testData: ChartData = {
-      name: 'test',
-      planets: [
-        { name: 'Sun', degree: 0 }, // 0° Aries (Luminary)
-        { name: 'Moon', degree: 6 }, // 6° Aries (Luminary) - 6° orb to Sun
-        { name: 'Mercury', degree: 90 }, // 0° Cancer (Personal) - square to Sun
-        { name: 'Venus', degree: 60 }, // 0° Gemini (Personal) - sextile to Sun
-        { name: 'Mars', degree: 120 }, // 0° Leo (Personal) - trine to Sun
-        { name: 'Jupiter', degree: 180 }, // 0° Libra (Social) - opposition to Sun
-        { name: 'Saturn', degree: 270 }, // 0° Capricorn (Social) - square to Sun
-        { name: 'Uranus', degree: 150 }, // 0° Virgo (Outer) - quincunx to Sun
-        { name: 'Neptune', degree: 30 }, // 0° Taurus (Outer) - semi-sextile to Sun
-        { name: 'Pluto', degree: 45 }, // 15° Taurus (Outer) - semi-square to Sun
-      ],
-    };
-
-    describe('OrbResolver class', () => {
-      test('resolves luminaries with larger orbs', () => {
-        const orbResolver = new OrbResolver(TRADITIONAL_ORB_CONFIG);
-
-        const sunMoonContext = {
-          chartType: 'natal' as const,
-          planetA: 'Sun',
-          planetB: 'Moon',
-          aspect: {
-            name: 'conjunction',
-            angle: 0,
-            orb: 5,
-            classification: AspectClassification.Major,
-          },
-        };
-
-        const orb = orbResolver.resolveOrb(sunMoonContext);
-        expect(orb).toBe(10); // Luminaries get 10° orb from TRADITIONAL_ORB_CONFIG
-      });
-
-      test('applies aspect classification multipliers', () => {
-        const orbResolver = new OrbResolver(MODERN_ORB_CONFIG);
-
-        const minorAspectContext = {
-          chartType: 'natal' as const,
-          planetA: 'Mercury',
-          planetB: 'Venus',
-          aspect: {
-            name: 'sextile',
-            angle: 60,
-            orb: 4,
-            classification: AspectClassification.Minor,
-          },
-        };
-
-        const orb = orbResolver.resolveOrb(minorAspectContext);
-        // Personal planets: 6° base, Minor aspects: 0.6 multiplier = 3.6°
-        expect(orb).toBe(3.6);
-      });
-
-      test('applies contextual orb multipliers for synastry', () => {
-        const orbResolver = new OrbResolver(TRADITIONAL_ORB_CONFIG);
-
-        const synastryContext = {
-          chartType: 'synastry' as const,
-          planetA: 'Sun',
-          planetB: 'Moon',
-          aspect: {
-            name: 'conjunction',
-            angle: 0,
-            orb: 5,
-            classification: AspectClassification.Major,
-          },
-        };
-
-        const orb = orbResolver.resolveOrb(synastryContext);
-        // Luminaries: 10° base, Synastry: 0.9 multiplier = 9.0°
-        expect(orb).toBe(9);
-      });
-
-      test('uses planet category mapping correctly', () => {
-        const orbResolver = new OrbResolver(TIGHT_ORB_CONFIG);
-
-        expect(orbResolver.getPlanetCategory('Sun')).toBe(
-          PlanetCategory.Luminaries
-        );
-        expect(orbResolver.getPlanetCategory('Mercury')).toBe(
-          PlanetCategory.Personal
-        );
-        expect(orbResolver.getPlanetCategory('Jupiter')).toBe(
-          PlanetCategory.Social
-        );
-        expect(orbResolver.getPlanetCategory('Uranus')).toBe(
-          PlanetCategory.Outer
-        );
-        expect(orbResolver.getPlanetCategory('Ascendant')).toBe(
-          PlanetCategory.Angles
-        );
-      });
-
-      test('handles cache correctly', () => {
-        const orbResolver = new OrbResolver(MODERN_ORB_CONFIG);
-
-        const context = {
-          chartType: 'natal' as const,
-          planetA: 'Sun',
-          planetB: 'Mars',
-          aspect: {
-            name: 'trine',
-            angle: 120,
-            orb: 6,
-            classification: AspectClassification.Major,
-          },
-        };
-
-        const orb1 = orbResolver.resolveOrb(context);
-        const orb2 = orbResolver.resolveOrb(context); // Should hit cache
-
-        expect(orb1).toBe(orb2);
-        expect(typeof orb1).toBe('number');
-      });
-    });
-
-    describe('Traditional Orb Configuration', () => {
-      test('uses traditional orbs with Sun-Moon conjunction', () => {
-        const result = chart2txt(testData, {
-          aspectDefinitions: TRADITIONAL_ASPECTS,
-          orbConfiguration: TRADITIONAL_ORB_CONFIG,
-          aspectCategories: [{ name: 'WIDE', maxOrb: 15 }], // Allow wide orbs for this test
-        });
-
-        // Sun-Moon should have 6° orb, which is within the 10° limit for luminaries
-        expect(result).toContain('Sun conjunction Moon');
-      });
-
-      test('gives outer planets smaller orbs', () => {
-        const result = chart2txt(testData, {
-          aspectDefinitions: [
-            {
-              name: 'semi-sextile',
-              angle: 30,
-              orb: 6,
-              classification: AspectClassification.Minor,
-            },
-          ],
-          orbConfiguration: TRADITIONAL_ORB_CONFIG,
-          aspectCategories: [{ name: 'ALL', maxOrb: 15 }], // Allow detection
-        });
-
-        // Neptune semi-sextile (30°) to Sun: outer planet gets 3° * 0.75 (minor) = 2.25° max orb
-        // But actual orb is 30° - 30° = 0°, so it should be included
-        expect(result).toContain('semi-sextile');
-      });
-    });
-
-    describe('Modern Orb Configuration', () => {
-      test('applies minor aspect orb reductions', () => {
-        const result = chart2txt(testData, {
-          aspectDefinitions: [
-            {
-              name: 'sextile',
-              angle: 60,
-              orb: 4,
-              classification: AspectClassification.Minor,
-            },
-          ],
-          orbConfiguration: MODERN_ORB_CONFIG,
-          aspectCategories: [{ name: 'ALL', maxOrb: 15 }], // Allow detection
-        });
-
-        // Venus sextile to Sun: Personal planet (6°) * Minor aspect (0.6) = 3.6° max orb
-        // Actual orb is 0°, so should be included
-        expect(result).toContain('sextile');
-      });
-    });
-
-    describe('Tight Orb Configuration', () => {
-      test('restricts aspects with tight orbs', () => {
-        const result = chart2txt(testData, {
-          aspectDefinitions: [
-            {
-              name: 'conjunction',
-              angle: 0,
-              orb: 5,
-              classification: AspectClassification.Major,
-            },
-          ],
-          orbConfiguration: TIGHT_ORB_CONFIG,
-          aspectCategories: [{ name: 'ALL', maxOrb: 15 }], // Allow detection
-        });
-
-        // Sun-Moon conjunction with 6° orb should be rejected (luminaries get max 5° in tight config)
-        expect(result).not.toContain('Sun conjunction Moon');
-      });
-
-      test('applies maximum orb constraints', () => {
-        const orbResolver = new OrbResolver(TIGHT_ORB_CONFIG);
-
-        const context = {
-          chartType: 'natal' as const,
-          planetA: 'Sun',
-          planetB: 'Moon',
-          aspect: {
-            name: 'conjunction',
-            angle: 0,
-            orb: 8,
-            classification: AspectClassification.Major,
-          },
-        };
-
-        const orb = orbResolver.resolveOrb(context);
-        // Should be capped at 5° maximum for major aspects in tight config
-        expect(orb).toBeLessThanOrEqual(5);
-      });
-    });
-
-    describe('Backwards Compatibility', () => {
-      test('works with legacy aspect definitions without orb configuration', () => {
-        const result = chart2txt(testData, {
-          aspectDefinitions: [
-            { name: 'conjunction', angle: 0, orb: 8 }, // Legacy format without classification
-          ],
-          aspectCategories: [{ name: 'LEGACY', maxOrb: 15 }], // Allow detection
-        });
-
-        // Should still work with traditional orb from aspect definition
-        expect(result).toContain('Sun conjunction Moon');
-      });
-
-      test('falls back to aspect orb when no orb configuration provided', () => {
-        const result = chart2txt(testData, {
-          aspectDefinitions: [
-            { name: 'trine', angle: 120, orb: 3 }, // Tight orb in aspect definition
-          ],
-          // No orbConfiguration provided
-        });
-
-        // Should use the 3° orb from the aspect definition, so Mars trine (0° orb) should be included
-        expect(result).toContain('trine');
-      });
-    });
-
-    describe('Synastry and Transit Contexts', () => {
-      test('applies different orbs for synastry', () => {
-        const chart1: ChartData = {
-          name: 'Person 1',
-          planets: [{ name: 'Sun', degree: 0 }],
-        };
-
-        const chart2: ChartData = {
-          name: 'Person 2',
-          planets: [{ name: 'Moon', degree: 7 }], // 7° orb
-        };
-
-        const result = chart2txt([chart1, chart2], {
-          aspectDefinitions: [
-            {
-              name: 'conjunction',
-              angle: 0,
-              orb: 5,
-              classification: AspectClassification.Major,
-            },
-          ],
-          orbConfiguration: TRADITIONAL_ORB_CONFIG, // Synastry multiplier: 0.9
-          aspectCategories: [{ name: 'SYNASTRY', maxOrb: 15 }], // Allow detection
-        });
-
-        // Luminaries: 10° * 0.9 = 9° max orb for synastry
-        // 7° actual orb should be within limits
-        expect(result).toContain('conjunction');
-      });
-
-      test('applies different orbs for transits', () => {
-        const natalChart: ChartData = {
-          name: 'Natal',
-          planets: [{ name: 'Sun', degree: 0 }],
-        };
-
-        const transitChart: ChartData = {
-          name: 'Transits',
-          planets: [{ name: 'Mars', degree: 8 }], // 8° orb
-          chartType: 'transit',
-        };
-
-        const result = chart2txt([natalChart, transitChart], {
-          aspectDefinitions: [
-            {
-              name: 'conjunction',
-              angle: 0,
-              orb: 5,
-              classification: AspectClassification.Major,
-            },
-          ],
-          orbConfiguration: TRADITIONAL_ORB_CONFIG, // Transit multiplier: 1.2
-          aspectCategories: [{ name: 'TRANSITS', maxOrb: 15 }], // Allow detection
-        });
-
-        // Luminaries: 10° * 1.2 = 12° max orb for transits
-        // 8° actual orb should be within limits
-        expect(result).toContain('conjunction');
-      });
-    });
-
-    describe('Error Handling and Edge Cases', () => {
-      test('handles unknown planet names gracefully', () => {
-        const orbResolver = new OrbResolver(MODERN_ORB_CONFIG);
-
-        const context = {
-          chartType: 'natal' as const,
-          planetA: 'UnknownPlanet',
-          planetB: 'AnotherUnknown',
-          aspect: {
-            name: 'conjunction',
-            angle: 0,
-            orb: 5,
-            classification: AspectClassification.Major,
-          },
-        };
-
-        const orb = orbResolver.resolveOrb(context);
-        // Should fall back to aspect orb
-        expect(orb).toBe(5);
-      });
-
-      test('handles missing orb configuration gracefully', () => {
-        const orbResolver = new OrbResolver(); // Empty configuration
-
-        const context = {
-          chartType: 'natal' as const,
-          planetA: 'Sun',
-          planetB: 'Moon',
-          aspect: {
-            name: 'conjunction',
-            angle: 0,
-            orb: 6,
-            classification: AspectClassification.Major,
-          },
-        };
-
-        const orb = orbResolver.resolveOrb(context);
-        // Should fall back to aspect orb
-        expect(orb).toBe(6);
-      });
-
-      test('applies global fallback orb when needed', () => {
-        const configWithFallback = {
-          ...TIGHT_ORB_CONFIG,
-          globalFallbackOrb: 2,
-        };
-
-        const orbResolver = new OrbResolver(configWithFallback);
-
-        const context = {
-          chartType: 'natal' as const,
-          planetA: 'UnknownPlanet',
-          planetB: 'AnotherUnknown',
-          aspect: { name: 'conjunction', angle: 0, orb: 0 }, // Invalid orb
-        };
-
-        const orb = orbResolver.resolveOrb(context);
-        expect(orb).toBeGreaterThan(0); // Should use fallback
-      });
-    });
-  });
-});
